+++ errors/response_errorer.go
@@ -1,0 +1,5 @@
+package errors
+
+type ResponseErrorer interface {
+	ResponseError() error
+}