+++ examples/app1/api/simple_test.go
@@ -16,9 +16,10 @@ func TestSimple(t *testing.T) {
 			Want:    json.ExpectBadRequest([]string{"missing data"}),
 		},
 		{
+			// NOTE: This is a generic error message so it's obscured by API middleware
 			Name:    "invalid-input",
 			Request: httptest.NewRequest("GET", "/simple?data=hey", nil),
-			Want:    json.ExpectInternalError(fmt.Errorf("invalid syntax")),
+			Want:    json.ExpectInternalError(fmt.Errorf("We have encountered an unexpected error.")),
 		},
 		{
 			Name:    "success",