+++ jsonapi/response_writer.go
@@ -4,6 +4,7 @@ import (
 	"bufio"
 	"errors"
 	"fmt"
+	api_errors "github.com/BSick7/go-api/errors"
 	"log"
 	"net"
 	"net/http"
@@ -26,6 +27,7 @@ var _ http.Hijacker = &ResponseWriter{}
 
 type ResponseWriter struct {
 	http.ResponseWriter
+	Obscurer   api_errors.Obscurer
 	start      time.Time
 	statusCode int
 }
@@ -79,11 +81,22 @@ func (w *ResponseWriter) SendError(err error) {
 			jaerr.Code = strconv.Itoa(isc.StatusCode())
 			jaerr.Status = http.StatusText(isc.StatusCode())
 		}
-		if payloader, ok := err.(ResponsePayloader); ok {
-			payload := payloader.Payload()
-			jaerr.Title = payload["title"].(string)
-			jaerr.Detail = payload["message"].(string)
+		var rerr api_errors.ResponseErrorer
+		if errors.As(err, &rerr) {
+			err = rerr.ResponseError()
 		}
+
+		// The interfaces defined in the errors package are not specific to jsonapi serialization
+		// This ResponsePayloader allows us to structure content before emitting in the jsonapi format
+		// Additionally, errors that aren't a ResponsePayloader are obscured (configurable through middleware)
+		payloader, ok := err.(ResponsePayloader)
+		if !ok {
+			payloader = w.Obscurer.Obscure(err)
+		}
+
+		payload := payloader.Payload()
+		jaerr.Title = payload["title"].(string)
+		jaerr.Detail = payload["message"].(string)
 	}
 	w.SendJsonApiError(jaerr)
 }